+++ jqm-all/jqm-api-client/jqm-api-client-core/src/main/java/com/enioka/jqm/client/shared/SimpleApiSecurity.java
@@ -111,7 +111,7 @@ public final class SimpleApiSecurity
                     Calendar expiration = Calendar.getInstance();
                     expiration.add(Calendar.DAY_OF_YEAR, 2);
 
-                    int id = RUser.create(cnx, login, hash, saltS, expiration, true, "administrator");
+                    long id = RUser.create(cnx, login, hash, saltS, expiration, true, "administrator");
                     currentUser = RUser.select_id(cnx, id);
                     user = currentUser;
