+++ jqm-all/jqm-integration-tests/src/test/java/com/enioka/jqm/integration/tests/ApiSimpleTest.java
@@ -90,10 +90,10 @@ public class ApiSimpleTest extends JqmBaseTest
 
         String result = res.body();
 
-        Integer jid = 0;
+        long jid = 0L;
         try
         {
-            jid = Integer.parseInt(result);
+            jid = Long.parseLong(result);
         }
         catch (Exception e)
         {
@@ -139,7 +139,7 @@ public class ApiSimpleTest extends JqmBaseTest
 
         String result = res.body();
 
-        Integer jid = 0;
+        long jid = 0;
         try
         {
             jid = Integer.parseInt(result);
